Suggested implementation for "alternative relations"
This implements lazy loading of related models as discussed in #70

The basic idea is that rather than using the type of the relation
attribute the name of the model class is provided to the
relationship traits and the actual class type is looked up
dynamically at run-time when it is actually required.

Because more is being done at run-time it doesn't perform some
compile time type checking and consequently some errors may be
somewhat less than awesome. There probably should be additional
tests for the failure modes (such as non-existent model class or
key name.)

It could also benefit from some refactoring as, for instance, the
type lookup is done in more than one place.

Hope that helps.