Allow creation of Red::Driver from an instance of DB
This allows the user to create the Red::Driver from an already
created and connected DBDish or DB::Pg object.

The rationale for doing this is to allow for the use of DB
specific functionality that may not be exposed by Red.

The Pg version I have tested with the code in #94 and it works.

Closes #94