$data structure fix for Postgres, so it behave consistently with SQLite
for SQLite, it re-fetches the row from the model by using LastInsertedRow
class, Then the `.row{@ids}` does the right thing all the time. But in
postgres, the $data will filled by `INSERT INTO table xxx RETURNING *` after
the RETURNING, we'll $data with column names with underscores.
So, in the pg case, We convert the $data structure according to the model, make
$data look like the same as fetched with LastInsertedRow class.