Don't do aliasing unless isn't done yet
This optimization is based on assumption that alias class name passed in
`:name` to method `alias` of `MetamodelX::Red::Model` unambigously
defines the class' typeobject. So, instead of re-aliasing on every
method call, `alias` records the class in a cache and re-uses when a
class of the same name is requested.

Hopefully, the approach is ok because caching is done per-model, so name
clashing should happen.