Fix create with more than one positional attribute (#468)
* Fix create with more than one positional attribute

When a model has 2+ positional relationships and 2 or more are defined
in the profile submitted to `.^create` then Red dies with an attempt to
find a positional accessor on `Any`. The bug is caused by scoping issue.
`FIRST` phaser doesn't create a new scope but operates withing the scope
of loop's block. Therefore any `my` declared within the phaser remains
uninitialized on 2nd repetition.

* Fix failing tests

`use v6` alone is sufficient for tests to pass. But I think it makes
sense to rename the modules with the recommended extensions.

* Use lib/ relative to test file location

The test file could be ran from anywhere, not only from the source root.