Fix for runtime application of role to model object
When a role is applied to an object with either 'does' or '.^mixin' the
compose method of the Meta object will be called to mix in any new
methods or attributes that the role provides, so the compose of the
MetamodelX::Red::Model must avoid creating duplicate accessors for the
model's columns.

Additionally when a role is applied with `does` the BUILD_LEAST_DERIVED
is called on the resulting composed object, resulting in the 'TWEAK' or
'BUILD' being called with no argument so the generated TWEAK added by
the MetamodelX::Red::Dirtable needs to deal with there already being
column data in the object and not overwriting that.

This fixes #478