Allow for aliases in Red::Schema
In a large application one might arrange the model classes in
multi-level packages to distinguish them from other classes in the
application (e.g. Org::AppName::Model::ModelName, ) and it might be
unwieldy (and difficult to remember,) to use the full name everywhere.

This provides for a Pair-wise `schema` so that an alias for the model
can be conveniently provided and used to retrieve the model.

Also provide a `model` method as an alternative to the `FALLBACK` as
using the fallback method may be confusing for a multi-level packaged
model.