Ensure use attr name not column name in new-from-data
If there was a column defined in a model where the DB column differs
from the attribute name, and the DB column name is the same as another
attribute in the model then `new-from-data` would pass the DB column
name to the constructor of the model class rather than the attribute
name that corresponds to that column which would lead either to a
failure if the attribute with that name was a relation (it would get a
plain value rather than a model object,) or the wrong attribute being
populated with the value.

This appears to only afflict Pg where the data from the insert with
`RETURNING *` is used to create the object returned by `create` whereas
for SQLite the object is loaded from the DB.

The test simulates the Pg behaviour by calling `new-from-data` directly.

Fixes #510