Don't emit "FOR UPDATE .." On SQLite
Rather than just ignore it like it does for other things it doesn't
implement, SQLite throws an error with 'FOR UPDATE' on a select
statement.

This just over-rides with an implementation that doesn't emit anything
if there is a '.for'.

The test just checks it doesn't explode with .skip-update, I've
confirmed the SQL generated for both Pg and SQLite visually.